requirements.txt
psycopg2-binary
python-dotenv
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
pandas